restructure redirectEnabled determination, add whitelisting of adresses
