+++ ext_conf_template.txt
@@ -7,5 +7,8 @@ testerIp =
 # cat=Features; type=string; label=LLL:EXT:td_mailredirect/Resources/Private/Language/ExtConfTemplate.xlf:feature.userAgent
 userAgent =
 
+# cat=Features; type=string; label=LLL:EXT:td_mailredirect/Resources/Private/Language/ExtConfTemplate.xlf:feature.whitelistedAddresses
+whitelistedAdresses =
+
 # cat=Features; type=string; label=LLL:EXT:td_mailredirect/Resources/Private/Language/ExtConfTemplate.xlf:feature.subjectTemplate
 subjectTemplate = [TEST] {subject}
